requirements.txt
# Core dependencies
python-telegram-bot>=20,<21
spotipy>=2.23.0

# Always latest (fast moving, frequent bugfixes)
yt-dlp
pydrive2

# Google auth stack (stable, safe to pin to major versions)
google-auth>=2.23.0,<3
google-auth-oauthlib>=1.1.0,<2
google-auth-httplib2>=0.2.0,<1

# Utilities
python-dotenv>=1.0.0
